Initial structure for the C++ hiveserver2 client.
The basic model: clients connect to the hiveserver2 server using
hs2client::Service::Connect. Sessions can then be started with
Service::OpenSession, and operations can be run on the session
with functions like Session::ExecuteStatement. The operation
handle provides functions for checking the operation status and
retrieving results.

The PIMPL idiom is used to hide Thrift from the public header files,
and a ThriftRPC struct containing the Thrift client object is passed
from the service into the session and operations to give them
access to RPC calls.

Currently, only fetching column oriented results is supported.
Fetched results are accessible via a ColumnarRowSet, and columns of
a given type can be accessed by their index. Column objects
expose pointers to the raw data returned by Thrift to give clients
access without unnecessary copying.

There is also a Status class that is used to indicate when RPCs
are unsuccessful, and a sample usage file to demonstrate how
a client can use the interface.

Change-Id: I7be82d5897ad6093116915b17924674c4ae508a0